+++ clid/const.py
@@ -251,4 +251,5 @@ PREF_DESC = {
     'vim_mode': 'Enable or disable Vim style keybindings ',
     'music_dir': 'Directory in which the app will search for mp3 files recursively ',
     'preview_format': 'Format in which a preview of the file under cursor will be shown ',
+    'use_regex_in_search': 'Enable or disable regular expressions when searching '
 }